optimize copyin out more, loadstore pass, pass all test for gpuratio = 1

+++ src/compiler/region.cpp
@@ -238,7 +238,7 @@ std::string petabricks::SimpleRegion::getIterationUpperBounds() const {
 
 std::string petabricks::SimpleRegion::getIterationMiddleEnd(std::string& middle) const {
   std::stringstream os;
-  for(int i = 0; i < maxCoord().size() - 1; ++i) {
+  for(size_t i = 0; i < maxCoord().size() - 1; ++i) {
     os << maxCoord()[i] << ",";
   }
   os << middle;
